refactor: move Request DTOs from presentation to application layer
