+++ backend/src/main/java/com/backend/application/dto/request/CreateComponentEntryRequest.java
@@ -1,4 +1,4 @@
-package com.backend.presentation.dto.request;
+package com.backend.application.dto.request;
 
 import com.backend.domain.enums.ComponentStatus;
 import jakarta.validation.constraints.NotNull;
