+++ backend/src/main/java/com/backend/application/dto/request/CreateEntryFieldRequest.java
@@ -1,4 +1,4 @@
-package com.backend.presentation.dto.request;
+package com.backend.application.dto.request;
 
 import com.backend.domain.enums.EntryFieldType;
 import jakarta.validation.constraints.NotBlank;
