+++ backend/src/main/java/com/backend/application/dto/request/CreateSectionRequest.java
@@ -1,10 +1,10 @@
-package com.backend.presentation.dto.request;
+package com.backend.application.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record UpdateSectionRequest(
-    @NotNull Long id,
+public record CreateSectionRequest(
+    @NotNull Long pageId,
     @Size(max = 50) String type,
     Integer displayOrder,
     String data) {
