+++ backend/src/main/java/com/backend/application/dto/request/PageCreateRequest.java
@@ -1,0 +1,23 @@
+package com.backend.application.dto.request;
+
+import com.backend.domain.enums.PageStatus;
+
+import jakarta.validation.constraints.Size;
+
+public record PageCreateRequest(
+    Long templateId,
+    PageStatus status,
+    @Size(max = 500, message = "validation.featured.image.size") String featuredImage,
+    @Size(max = 255, message = "validation.style.classes.size") String styleClasses,
+    Integer sortOrder) {
+
+  public PageCreateRequest {
+    if (featuredImage != null) {
+      featuredImage = featuredImage.trim();
+    }
+    if (styleClasses != null) {
+      styleClasses = styleClasses.trim();
+    }
+  }
+}
+
