+++ backend/src/main/java/com/backend/application/dto/request/PagePublishRequest.java
@@ -1,0 +1,19 @@
+package com.backend.application.dto.request;
+
+import java.time.LocalDateTime;
+
+import jakarta.validation.constraints.NotNull;
+
+public record PagePublishRequest(
+    @NotNull(message = "validation.tenant.id.required") Long tenantId,
+    LocalDateTime scheduledAt) {
+
+  public boolean isImmediatePublish() {
+    return scheduledAt == null;
+  }
+
+  public boolean isScheduledPublish() {
+    return scheduledAt != null;
+  }
+}
+
