+++ backend/src/main/java/com/backend/application/dto/request/UpdateSectionRequest.java
@@ -1,10 +1,10 @@
-package com.backend.presentation.dto.request;
+package com.backend.application.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record CreateSectionRequest(
-    @NotNull Long pageId,
+public record UpdateSectionRequest(
+    @NotNull Long id,
     @Size(max = 50) String type,
     Integer displayOrder,
     String data) {
