+++ backend/src/main/java/com/backend/application/service/PageI18nService.java
@@ -1,8 +1,8 @@
 package com.backend.application.service;
 
 import com.backend.domain.enums.Language;
-import com.backend.presentation.dto.request.PageI18nRequest;
-import com.backend.presentation.dto.request.PagePublishRequest;
+import com.backend.application.dto.request.PageI18nRequest;
+import com.backend.application.dto.request.PagePublishRequest;
 import com.backend.presentation.dto.response.PageI18nResponse;
 import java.util.List;
