+++ backend/src/main/java/com/backend/application/service/PageService.java
@@ -1,6 +1,6 @@
 package com.backend.application.service;
 
-import com.backend.presentation.dto.request.PageCreateRequest;
+import com.backend.application.dto.request.PageCreateRequest;
 import com.backend.presentation.dto.response.PageListResponse;
 import com.backend.presentation.dto.response.PageResponse;
 import com.backend.presentation.dto.response.PageDetailResponse;
@@ -8,10 +8,16 @@ import java.util.List;
 
 public interface PageService {
     PageResponse createPage(PageCreateRequest request, Long userId);
+
     PageResponse getPageById(Long id);
+
     PageDetailResponse getPageWithI18n(Long id);
+
     List<PageResponse> getAllPages();
+
     List<PageListResponse> getAllPagesWithTranslations();
+
     PageResponse updatePage(Long id, PageCreateRequest request, Long userId);
+
     void deletePage(Long id);
 }
