+++ backend/src/main/java/com/backend/application/service/SiteService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Optional;
 
 import com.backend.domain.enums.Language;
-import com.backend.presentation.dto.request.CreateSiteRequest;
-import com.backend.presentation.dto.request.UpdateSiteRequest;
+import com.backend.application.dto.request.CreateSiteRequest;
+import com.backend.application.dto.request.UpdateSiteRequest;
 import com.backend.presentation.dto.response.SiteResponse;
 
 public interface SiteService {
