+++ backend/src/main/java/com/backend/application/service/SiteServiceImpl.java
@@ -7,8 +7,8 @@ import com.backend.domain.exception.TenantNotFoundException;
 import com.backend.domain.repository.SiteRepository;
 import com.backend.domain.repository.TenantRepository;
 import com.backend.presentation.dto.mapper.SiteMapper;
-import com.backend.presentation.dto.request.CreateSiteRequest;
-import com.backend.presentation.dto.request.UpdateSiteRequest;
+import com.backend.application.dto.request.CreateSiteRequest;
+import com.backend.application.dto.request.UpdateSiteRequest;
 import com.backend.presentation.dto.response.SiteResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
