+++ backend/src/main/java/com/backend/presentation/controller/ComponentEntryController.java
@@ -6,9 +6,9 @@ import com.backend.domain.entity.ComponentEntry;
 import com.backend.domain.entity.ComponentEntryI18n;
 import com.backend.domain.enums.ComponentStatus;
 import com.backend.domain.enums.Language;
-import com.backend.presentation.dto.request.CreateComponentEntryRequest;
+import com.backend.application.dto.request.CreateComponentEntryRequest;
 import com.backend.presentation.dto.request.EntryI18nRequest;
-import com.backend.presentation.dto.request.UpdateComponentEntryRequest;
+import com.backend.application.dto.request.UpdateComponentEntryRequest;
 import com.backend.presentation.dto.response.ComponentEntryI18nResponse;
 import com.backend.presentation.dto.response.ComponentEntryResponse;
 import com.backend.shared.common.ApiResponse;
