+++ backend/src/main/java/com/backend/presentation/controller/TenantController.java
@@ -7,8 +7,8 @@ import com.backend.application.usecase.CreateTenantUseCase;
 import com.backend.application.usecase.GenerateTenantAdminUserUseCase;
 import com.backend.domain.enums.Language;
 import com.backend.domain.enums.TenantStatus;
-import com.backend.presentation.dto.request.CreateTenantRequest;
-import com.backend.presentation.dto.request.UpdateTenantRequest;
+import com.backend.application.dto.request.CreateTenantRequest;
+import com.backend.application.dto.request.UpdateTenantRequest;
 import com.backend.presentation.dto.response.TenantDetailResponse;
 import com.backend.presentation.dto.response.TenantListResponse;
 import com.backend.shared.common.ApiResponse;
