+++ backend/src/main/java/com/backend/presentation/dto/mapper/SiteMapper.java
@@ -2,8 +2,8 @@ package com.backend.presentation.dto.mapper;
 
 import com.backend.domain.entity.Site;
 import com.backend.domain.enums.Language;
-import com.backend.presentation.dto.request.CreateSiteRequest;
-import com.backend.presentation.dto.request.UpdateSiteRequest;
+import com.backend.application.dto.request.CreateSiteRequest;
+import com.backend.application.dto.request.UpdateSiteRequest;
 import com.backend.presentation.dto.response.SiteResponse;
 import org.springframework.stereotype.Component;
