--- backend/src/main/java/com/backend/presentation/dto/request/CreateBlockRequest.java
@@ -1,11 +1,0 @@
-package com.backend.presentation.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-public record CreateBlockRequest(
-    @NotNull Long sectionId,
-    @Size(max = 50) String type,
-    Integer displayOrder,
-    String data) {
-}
