--- backend/src/main/java/com/backend/presentation/dto/request/PagePublishRequest.java
@@ -1,19 +1,0 @@
-package com.backend.presentation.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDateTime;
-
-public record PagePublishRequest(
-
-        @NotNull(message = "validation.tenant.id.required") Long tenantId,
-
-        LocalDateTime scheduledAt) {
-
-    public boolean isImmediatePublish() {
-        return scheduledAt == null;
-    }
-
-    public boolean isScheduledPublish() {
-        return scheduledAt != null;
-    }
-}
