--- backend/src/main/java/com/backend/presentation/dto/request/ReorderRequest.java
@@ -1,9 +1,0 @@
-package com.backend.presentation.dto.request;
-
-import java.util.List;
-
-import jakarta.validation.constraints.NotEmpty;
-
-public record ReorderRequest<T>(
-    @NotEmpty(message = "Items list cannot be empty") List<T> items) {
-}
